# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorWrapper:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.SuccessWrapper:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UserDetailRequest:
    properties:
      address:
        type: string
      blood_type:
        enum:
        - A
        - B
        - AB
        - O
        type: string
      date_of_birth:
        type: string
      full_name:
        type: string
      gender:
        enum:
        - L
        - P
        type: string
      is_active_donor:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      phone_number:
        maxLength: 15
        minLength: 10
        type: string
      rhesus:
        enum:
        - +
        - '-'
        type: string
      weight:
        type: number
    required:
    - address
    - blood_type
    - date_of_birth
    - full_name
    - gender
    - is_active_donor
    - phone_number
    - rhesus
    type: object
  dto.UserRequest:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        type: string
    required:
    - password
    type: object
info:
  contact:
    email: support@donor-darah.duckdns.org
    name: Ulil Amry
    url: https://donor-darah.duckdns.org
  description: API untuk mengelola donor darah, stok darah, dan informasi pendonor
    dengan autentikasi OAuth.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://donor-darah.duckdns.org/terms
  title: Donor App API
  version: "1.0"
paths:
...
  /profile:
    get:
      description: Mengambil profil dasar dan detail dari pengguna yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: Profil berhasil diambil
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "401":
          description: Tidak terautentikasi
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "404":
          description: Profil tidak ditemukan
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Get current user's profile
      tags:
      - Profile
  /profile/detail:
    get:
      description: Mengambil profil detail dari pengguna yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: Detail profil berhasil diambil
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "401":
          description: Tidak terautentikasi
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "404":
          description: Detail profil tidak ditemukan
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Get my user detail
      tags:
      - Profile Details
  /profile/detail/update:
    put:
      consumes:
      - application/json
      description: Memperbarui profil detail dari pengguna yang sedang login
      parameters:
      - description: Data Detail Profil yang Diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Detail profil berhasil diperbarui
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "401":
          description: Tidak terautentikasi
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "500":
          description: Terjadi kesalahan internal
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Update my user detail
      tags:
      - Profile Details
  /profile/details:
    post:
      consumes:
      - application/json
      description: Membuat profil detail (NIK, alamat, dll.) untuk pengguna yang sedang
        login. Hanya bisa dibuat sekali.
      parameters:
      - description: Data Detail Profil
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserDetailRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Detail profil berhasil dibuat
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "401":
          description: Tidak terautentikasi
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "500":
          description: Terjadi kesalahan internal
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Create my user detail
      tags:
      - Profile Details
  /profile/update:
    put:
      consumes:
      - application/json
      description: Memperbarui informasi dasar (nama, email) dari pengguna yang sedang
        login
      parameters:
      - description: Data Profil yang Diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profil berhasil diperbarui
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "401":
          description: Tidak terautentikasi
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "500":
          description: Terjadi kesalahan internal
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Update current user's profile
      tags:
      - Profile
...
  /users:
    get:
      description: Mengambil daftar semua lokasi dengan paginasi
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah item per halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil daftar lokasi
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "500":
          description: Terjadi kesalahan internal
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Membuat profil detail (NIK, alamat, dll.) untuk pengguna yang sedang
        login. Hanya bisa dibuat sekali.
      parameters:
      - description: Data Detail Profil
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserDetailRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Detail profil berhasil dibuat
          schema:
            $ref: '#/definitions/dto.SuccessWrapper'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "401":
          description: Tidak terautentikasi
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
        "500":
          description: Terjadi kesalahan internal
          schema:
            $ref: '#/definitions/dto.ErrorWrapper'
      security:
      - BearerAuth: []
      summary: Create all user data
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"